Do not over-allocate. Use a size_t for the max size.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@34487 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_entry.h
@@ -55,7 +55,7 @@ struct Etk_Entry
    Ecore_IMF_Context *imf_context;
 
    char *text;
-   int text_limit;
+   size_t text_limit;
 
    Etk_Color highlight_color;
    int image_interspace;
@@ -76,8 +76,8 @@ Etk_Widget *etk_entry_new(void);
 
 void        etk_entry_text_set(Etk_Entry *entry, const char *text);
 const char *etk_entry_text_get(Etk_Entry *entry);
-void        etk_entry_text_limit_set(Etk_Entry *entry, int limit);
-int         etk_entry_text_limit_get(Etk_Entry *entry);
+void        etk_entry_text_limit_set(Etk_Entry *entry, size_t limit);
+size_t      etk_entry_text_limit_get(Etk_Entry *entry);
 void        etk_entry_clear(Etk_Entry *entry);
 void        etk_entry_image_set(Etk_Entry *entry, Etk_Entry_Image_Position position, Etk_Image *image);
 Etk_Image  *etk_entry_image_get(Etk_Entry *entry, Etk_Entry_Image_Position position);
